Fully worked AlertService
